+++ app/ShuffledItem.php
@@ -7,6 +7,8 @@ use Astrotomic\Translatable\Contracts\Translatable as TranslatableContract;
 use Astrotomic\Translatable\Translatable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Cache;
+use Mcamara\LaravelLocalization\Facades\LaravelLocalization;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\InteractsWithMedia;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
@@ -18,6 +20,15 @@ class ShuffledItem extends Model implements HasMedia, TranslatableContract
     use InteractsWithMedia;
     use Publishable;
 
+    protected static function booted()
+    {
+        static::saved(function () {
+            foreach (LaravelLocalization::getSupportedLanguagesKeys() as $locale) {
+                Cache::forget("home.shuffled-items.$locale");
+            }
+        });
+    }
+
     public $translatedAttributes = ['filters'];
 
     protected $casts = [
@@ -25,6 +36,32 @@ class ShuffledItem extends Model implements HasMedia, TranslatableContract
     ];
 
     protected $fillable = ['is_published', 'item_id', 'crop'];
+
+    public function toShuffleOrchestratorItem()
+    {
+        return [
+            'id' => $this->id,
+            'authorLinks' => $this->author_links,
+            'title' => $this->title,
+            'url' => route('dielo', $this->item),
+            'datingFormatted' => $this->dating_formatted,
+            'filters' => collect($this->filters)->map(
+                fn($f) => [
+                    'url' => $f['url'],
+                    'attributes' => collect($f['attributes'])->map(
+                        fn($a) => [
+                            'label' => trans('item.' . $a['name']),
+                            'value' => $a['label'],
+                        ]
+                    ),
+                ]
+            ),
+            'img' => [
+                'src' => $this->getFirstMedia('image')->getUrl(),
+                'srcset' => $this->getFirstMedia('image')->getSrcset(),
+            ],
+        ];
+    }
 
     public function item()
     {
@@ -47,7 +84,7 @@ class ShuffledItem extends Model implements HasMedia, TranslatableContract
     {
         $iipParams = [
             'FIF' => $this->item->images()->first()->iipimg_url,
-            'WID' => 1600, // Max size
+            'WID' => 1920, // Max size
             'RGN' => join(',', [
                 $this->crop['x'],
                 $this->crop['y'],