+++ resources/views/home/index.blade.php
@@ -25,37 +25,119 @@
 
 @section('content')
     <div class="tailwind-rules">
-        @if ($featuredPiece)
-            <div class="tw-relative">
-                {{ $featuredPiece->image->img()->attributes(['class' => 'tw-object-cover tw-absolute tw-h-full tw-w-full', 'width' => null, 'height' => null]) }}
-                <div
-                    class="tw-absolute tw-inset-0 tw-bg-gradient-to-r tw-from-black/50 tw-to-black/20 md:tw-to-black/0">
-                </div>
+        @if ($shuffledItems->count() > 0)
+            <home.shuffle-orchestrator v-bind:items="{{ Js::from($shuffledItems) }}"
+                v-slot="orchestrator">
+                <div class="tw-relative tw-overflow-hidden">
+                    <img v-bind:class="['tw-absolute tw-h-full tw-w-full tw-object-cover tw-transition-all tw-scale-[1.005] tw-duration-700 tw-ease-in', {'tw-blur tw-scale-100': orchestrator.isShuffling }]"
+                        v-bind:src="orchestrator.item.img.src"
+                        v-bind:srcset="orchestrator.item.img.srcset"
+                        sizes="(max-width: 480px) 250vw, (max-width: 640px) 150vw, 100vw" />
 
-                <featured-piece-click-tracker v-bind:id="{{ $featuredPiece->id }}"
-                    url="{{ $featuredPiece->url }}" v-slot="{ track }">
-                    <div
-                        class="tw-container tw-relative tw-mx-auto tw-max-w-screen-xl tw-py-8 tw-px-6 tw-text-white md:tw-py-20">
-                        <h2 class="tw-mt-36 tw-font-semibold tw-drop-shadow md:tw-mt-48 md:tw-text-lg">
-                            {{ trans('home.featured_piece.title') }}
-                        </h2>
-                        <h3 class="tw-mt-4 tw-text-3xl tw-font-semibold tw-drop-shadow md:tw-text-6xl">
-                            <a href="{{ $featuredPiece->url }}" v-on:click.once.prevent="track">
-                                {{ $featuredPiece->title }}
-                            </a>
-                        </h3>
-                        <div
-                            class="tw-prose-invert prose-p:tw-mb-4 tw-mt-5 tw-max-w-lg tw-font-serif tw-leading-relaxed tw-drop-shadow md:tw-text-xl">
-                            {!! $featuredPiece->excerpt !!}
-                        </div>
-                        <a href="{{ $featuredPiece->url }}" v-on:click.once.prevent="track"
-                            class="tw-mt-3 tw-inline-block tw-border tw-border-gray-300 tw-px-4 tw-py-2 tw-text-sm tw-transition tw-duration-300 hover:tw-border-gray-400 hover:tw-bg-white hover:tw-text-gray-800 md:tw-mt-6">
-                            {{ $featuredPiece->is_collection? trans('home.featured_piece.button_collection'): trans('home.featured_piece.button_article') }}
-                            <i class="fa icon-arrow-right tw-ml-2"></i>
-                        </a>
+                    <img class="tw-invisible tw-absolute tw-h-full tw-w-full tw-scale-[1.005] tw-object-cover"
+                        v-bind:src="orchestrator.nextImg.src"
+                        v-bind:srcset="orchestrator.nextImg.srcset"
+                        sizes="(max-width: 480px) 250vw, (max-width: 640px) 150vw, 100vw" />
+
+                    {{-- Partial bottom overlay (for readability) --}}
+                    <div v-cloak
+                        class="tw-absolute tw-inset-x-0 tw-bottom-0 tw-h-1/2 tw-bg-gradient-to-b tw-from-black/0 tw-to-black/60 md:tw-h-1/3 md:tw-to-black/50">
                     </div>
-                </featured-piece-click-tracker>
-            </div>
+
+                    {{-- Full overlay for transitions --}}
+                    <div v-cloak
+                        v-bind:class="['tw-absolute tw-inset-0 tw-transition-all tw-bg-black tw-duration-500', orchestrator.isShuffling ? 'tw-opacity-50' : 'tw-opacity-0']">
+                    </div>
+                    <div
+                        class="tw-relative tw-flex tw-flex-col tw-items-center tw-p-6 tw-text-white md:tw-p-8">
+                        <h1
+                            class="tw-mt-20 tw-text-center tw-text-3xl tw-drop-shadow-[0_1px_2px_rgba(0,0,0,0.7)] md:tw-mt-40 md:tw-text-6xl md:tw-drop-shadow-[0_4px_3px_rgba(0,0,0,0.5)]">
+                            {{ trans('home.shuffled_item.tagline') }}
+                        </h1>
+                        <div class="tw-max-w-5xl tw-text-center">
+                            <p
+                                class="tw-mt-2 tw-drop-shadow-[0_1px_2px_rgba(0,0,0,0.7)] md:tw-mt-6 md:tw-text-2xl">
+                                {!! trans('home.shuffled_item.subtitle') !!}
+                            </p>
+                        </div>
+
+                        <div class="tw-mt-6 tw-self-stretch md:tw-mt-10" v-cloak>
+                            <div
+                                class="tw-mx-auto tw-flex tw-w-60 tw-flex-col md:tw-w-auto md:tw-max-w-3xl">
+                                <div class="tw-flex tw-flex-col tw-items-stretch">
+                                    <div class="tw-justify-items-stretch md:tw-flex">
+                                        <div
+                                            class="tw-grid tw-grow tw-grid-cols-1 tw-gap-2.5 tw-bg-black/60 tw-p-4 tw-shadow md:tw-grid-cols-3">
+
+                                            <div v-cloak
+                                                v-for="filterAttribute, filterAttributeIndex in orchestrator.filter.attributes"
+                                                v-bind:key="filterAttributeIndex">
+
+                                                <home.transition-in-place
+                                                    v-bind:transition-key="filterAttribute.label + filterAttribute.value">
+                                                    <div>
+                                                        <div class="tw-text-xs tw-text-white/40">
+                                                            @{{ filterAttribute.label }}
+                                                        </div>
+                                                        <div
+                                                            v-bind:class="['tw-whitespace-nowrap tw-transition-opacity tw-text-sm md:tw-text-base tw-text-white', {'tw-opacity-40': orchestrator.isShuffling }]">
+                                                            @{{ filterAttribute.value }}
+                                                        </div>
+                                                    </div>
+                                                </home.transition-in-place>
+                                            </div>
+                                        </div>
+                                        <button v-on:click="orchestrator.shuffle"
+                                            class="tw-group tw-w-full tw-basis-0 tw-bg-sky-300 tw-px-4 tw-py-2 tw-text-center tw-text-xs tw-text-black tw-transition-colors hover:tw-bg-sky-400 md:tw-px-6 md:tw-text-sm">
+                                            <i
+                                                class="fa fa-repeat tw-mr-2 tw--ml-4 tw-transition-transform group-hover:tw-rotate-45 md:tw-mx-0"></i>
+                                            {{ trans('home.shuffled_item.button_shuffle') }}
+                                        </button>
+                                    </div>
+
+                                </div>
+
+                                <x-home.button v-bind:href="orchestrator.filter.url"
+                                    v-bind:class="['tw-mt-6 tw-self-stretch tw-bg-white/10 tw-text-center md:tw-hidden', {'tw-opacity-40 tw-pointer-events-none': orchestrator.isShuffling}]">
+                                    {{ trans('home.shuffled_item.more_like_this') }}
+                                </x-home.button>
+                            </div>
+                        </div>
+
+                        <div class="tw-mt-10 tw-grid tw-self-stretch md:tw-mt-20 md:tw-grid-cols-3">
+                            <div
+                                class="tw-col-start-2 tw-hidden tw-text-center md:tw-block md:tw-pb-24">
+                                <x-home.button v-bind:href="orchestrator.filter.url"
+                                    v-bind:class="['tw-self-stretch tw-bg-white/10 tw-text-center disabled:tw-opacity-40', {'tw-opacity-40 tw-pointer-events-none': orchestrator.isShuffling}]">
+                                    {{ trans('home.shuffled_item.more_like_this') }}
+                                </x-home.button>
+                            </div>
+                            <div
+                                v-bind:class="['tw-text-white/60 tw-flex tw-flex-col tw-items-center tw-text-xs md:tw-self-end md:tw-items-end md:tw-text-sm tw-transition-opacity tw-duration-500', {'tw-opacity-0 tw-pointer-events-none': orchestrator.isShuffling}]">
+                                <div
+                                    class="tw-hidden tw-flex-col tw-items-end tw-justify-end tw-gap-x-1 tw-text-right md:tw-flex">
+                                    <span>@{{ orchestrator.item.authors }}</span>
+                                    <strong>@{{ orchestrator.item.title }}</strong>
+                                    <span>@{{ orchestrator.item.dating }}</span>
+                                </div>
+                                <div class="text-center tw-justify-center md:tw-hidden">
+                                    <span>@{{ orchestrator.item.authors }},</span>
+                                    <strong>@{{ orchestrator.item.title }}</strong>,
+                                    <span>@{{ orchestrator.item.dating }}</span>
+                                </div>
+
+                                <div class="tw-mt-2 tw-text-center">
+                                    <a v-bind:href="orchestrator.item.url"
+                                        class="tw-underline tw-decoration-2 tw-underline-offset-2 tw-transition-colors hover:tw-text-white">
+                                        {{ trans('home.shuffled_item.go_to_item') }}
+                                    </a>
+                                </div>
+                            </div>
+
+                        </div>
+                    </div>
+                </div>
+            </home.shuffle-orchestrator>
         @endif
 
         {{-- Counts blurb --}}
@@ -75,6 +157,40 @@
             </div>
         </div>
 
+        @if ($featuredPiece)
+            <div class="tw-relative">
+                {{ $featuredPiece->image->img()->attributes(['class' => 'tw-object-cover tw-absolute tw-h-full tw-w-full', 'width' => null, 'height' => null]) }}
+                <div
+                    class="tw-absolute tw-inset-0 tw-bg-gradient-to-r tw-from-black/50 tw-to-black/20 md:tw-to-black/0">
+                </div>
+
+                <featured-piece-click-tracker v-bind:id="{{ $featuredPiece->id }}"
+                    url="{{ $featuredPiece->url }}" v-slot="{ track }">
+                    <div
+                        class="tw-container tw-relative tw-mx-auto tw-flex tw-max-w-screen-xl tw-flex-col tw-items-start tw-py-8 tw-px-6 tw-text-white md:tw-py-20">
+                        <a href="{{ $featuredPiece->url }}" v-on:click.once.prevent="track"
+                            class="tw-absolute tw-inset-0"></a>
+
+                        <h2 class="tw-mt-36 tw-font-semibold tw-drop-shadow md:tw-mt-48 md:tw-text-lg">
+                            {{ trans('home.featured_piece.title') }}
+                        </h2>
+                        <h3 class="tw-mt-4 tw-text-3xl tw-font-semibold tw-drop-shadow md:tw-text-6xl">
+                            {{ $featuredPiece->title }}
+                        </h3>
+                        <div
+                            class="tw-prose-invert tw-mt-5 tw-max-w-lg tw-font-serif tw-leading-relaxed tw-drop-shadow prose-p:tw-mb-4 md:tw-text-xl">
+                            {!! $featuredPiece->excerpt !!}
+                        </div>
+                        <x-home.button href="{{ $featuredPiece->url }}"
+                            v-on:click.once.prevent="track"
+                            class="tw-relative tw-mt-3 tw-bg-white/10 md:tw-mt-6">
+                            {{ $featuredPiece->is_collection? trans('home.featured_piece.button_collection'): trans('home.featured_piece.button_article') }}
+                        </x-home.button>
+                    </div>
+                </featured-piece-click-tracker>
+            </div>
+        @endif
+
         @if ($featuredArtwork)
             <div class="tw-bg-gray-800 tw-text-white">
                 <div
@@ -82,8 +198,7 @@
                     <div class="tw-text-center lg:tw-order-1">
                         <a href="{{ route('dielo', ['id' => $featuredArtwork->item->id]) }}"
                             class="tw-inline-block">
-                            <x-item_image :id="$featuredArtwork->item->id"
-                                alt="{{ $featuredArtwork->title }}"
+                            <x-item_image :id="$featuredArtwork->item->id" alt="{{ $featuredArtwork->title }}"
                                 class="tw-max-h-80 lg:tw-max-h-[32rem]" />
                         </a>
                     </div>
@@ -114,14 +229,13 @@
                             @endforeach
                         </div>
                         <div
-                            class="tw-prose-invert prose-p:tw-mb-4 tw-mt-4 tw-hidden tw-font-serif tw-text-xl tw-leading-relaxed lg:tw-block">
+                            class="tw-prose-invert tw-mt-4 tw-hidden tw-font-serif tw-text-xl tw-leading-relaxed prose-p:tw-mb-4 lg:tw-block">
                             {!! $featuredArtwork->description !!}
                         </div>
-                        <a href="{{ route('dielo', ['id' => $featuredArtwork->item->id]) }}"
-                            class="tw-mt-6 tw-inline-block tw-border tw-border-gray-300 tw-px-4 tw-py-2 tw-text-sm tw-transition tw-duration-300 hover:tw-border-gray-400 hover:tw-bg-white hover:tw-text-gray-800">
-                            {{ trans('home.featured_artwork.button') }}
-                            <i class="fa icon-arrow-right tw-ml-2"></i>
-                        </a>
+                        <x-home.button
+                            href="{{ route('dielo', ['id' => $featuredArtwork->item->id]) }}"
+                            class="tw-mt-6">{{ trans('home.featured_artwork.button') }}
+                        </x-home.button>
                     </div>
                 </div>
             </div>
@@ -139,7 +253,7 @@
                         @foreach ([trans('home.latest_content.collections.tab'), trans('home.latest_content.articles.tab')] as $tab)
                             <tab v-slot="{ active }">
                                 <button
-                                    :class="['tw-transition-colors tw-font-semibold tw-text-2xl md:tw-text-4xl',
+                                    v-bind:class="['tw-transition-colors tw-font-semibold tw-text-2xl md:tw-text-4xl',
                                         !active && 'tw-text-gray-300 hover:tw-underline tw-underline-offset-[5px] md:tw-underline-offset-8 tw-decoration-[3px]'
                                     ]">{{ $tab }}</button>
                             </tab>
@@ -279,11 +393,11 @@
                                     {{ $featuredAuthor->death_place }}
                                 @endif
                             </div>
-                            <a href="{{ route('frontend.author.detail', $featuredAuthor) }}"
-                                class="tw-mt-6 tw-inline-block tw-border tw-border-gray-300 tw-px-4 tw-py-2 tw-text-sm tw-transition tw-duration-300 hover:tw-border-gray-400 hover:tw-bg-white hover:tw-text-gray-800">
+                            <x-home.button
+                                href="{{ route('frontend.author.detail', $featuredAuthor) }}"
+                                class="tw-mt-6">
                                 {!! trans('home.featured_author.button_author', ['count' => $featuredAuthor->items_count]) !!}
-                                <i class="fa icon-arrow-right tw-ml-2"></i>
-                            </a>
+                            </x-home.button>
                             <br />
                             <a href="{{ route('frontend.author.index') }}"
                                 class="tw-mt-6 tw-hidden tw-cursor-pointer tw-text-sm tw-underline tw-decoration-gray-300 tw-underline-offset-4 tw-transition-colors hover:tw-decoration-current lg:tw-inline-block">