+++ CHANGELOG.md
@@ -3,11 +3,18 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+## [2.55.0] - 2022-06-03
+### Added
+- front-end components for "shuffle" items
+
+### Changed
+- whole featured piece home section to be clickable
+
 ## [2.54.0] - 2022-05-19
 ### Changed
 - Removed PNP from downloadable items
 - Information page with updated data
--
+
 ## [2.53.0] - 2022-05-13
 ### Changed
 - temporarily removed print reproductions